Fix infinite searching for `.swift-format` file on Linux
Due to https://github.com/swiftlang/swift-foundation/issues/980, the we may end up with a `path == ""` instead of `/`.

We didn’t catch this in the test because we only end up with an empty path when searching for a `.swift-format` file from a `.swift-file` that was not at the root of a directory.