PrettyPrinterPerformance PrettyPrinterPerformance PrettyPrinterPerformance PrettyPrinterPerformance Optimized the PrettyPrinter for #894
Worked to get the perfomance to be closer to where we were before the changes in #883. This code should be only about 1.5% slower rather than 7% slower.

Using a lazy filter as `count(where:_)` isn't avaliable < Swift 6.0.

One forward loop and using the UTF8 view makes this faster than the original code pre-#883.