Add dump-effective-configuration subcommand
closes #667

Implement the subcommand `dump-effective-configuration`, which dumps
the configuration that would be used if `swift-format` was executed
from the current working directory (cwd), incorporating configuration
files found in the cwd or its parents, or input from the
`--configuration` option.

This helps when composing a configuration or with configuration
debugging/verification activities.