Do not measure perform performance in CI
When executing in Swift CI, run the block to make sure it doesn't hit any assertions because we don't look at performance numbers in CI and CI nodes can have variable performance characteristics if they are not bare-metal.

Anywhere else, run XCTest's `measure` function to measure the performance of the block.

Fixes #939
rdar://145139964